Add weight convert, weight cast and weight prepack for fp16 (#1134)
* add weight convert and weight prepack for fp16

* fix UT

* update ideep for fp16

* add fp16 support check

* add checks for weight prepack

* reuse existing methods

+++ tests/cpu/test_ipex_optimize.py
@@ -121,10 +121,10 @@ class TestOptimizeCases(TestCase):
                 super(Conv, self).__init__()
                 self.conv = torch.nn.Conv2d(3, 64, kernel_size=(7, 7), stride=(2, 2), padding=(3, 3), bias=False)
 
-        def forward(self, x):
-            return self.conv(x)
+            def forward(self, x):
+                return self.conv(x)
 
-        model = Conv().half()
+        model = Conv().double()
         with self.assertWarnsRegex(UserWarning,
                                    "WARNING: Can't convert model's parameters dtype"):
             optimized_model = ipex.optimize(model.eval(), dtype=torch.bfloat16)
